==> 4_evaluate.ipynb <==
import pickle
import pandas as pd
import numpy as np

cached = {}

def load(data_name, fold, group_f, rating, ll, nth, topk):
    path = f"data_splits/{data_name}/{fold}/CEs/stats_{group_f}_{rating}_{ll}_{nth}_{topk}.pickle"
    if path in cached:
        return cached[path]
    try:
        with open(path, "rb") as f:
            allstats = pickle.load(f)
    except FileNotFoundError:
        return {}
    cached[path] = allstats
    return allstats


def get_val_for_confs(valnames, dataname=["yelp", "netflix", "amazon"], aggf=["sum", "mean", "disjunction"], rating=["rating", "binary"], ll=["median", "0.1", "no_spn"], ordering=["score", "nth"], topk=[5, 10]):
    all_vals, all_fails = [], 0
    if type(valnames) != list:
        valnames = [valnames]
    for dname in dataname:
        for af in aggf:
            for rat in rating:
                for l in ll:
                    for ord in ordering:
                        for tk in topk:
                            vals, fails = get_val_for_conf(valnames, dname, af, rat, l, ord, tk)
                            all_vals += vals
                            all_fails += fails
    return all_vals, all_fails

def get_val_for_conf(valnames, dataname, aggf, rating, ll, ordering, topk):
    vals = []
    fails = 0
    for fold in range(3):
        data = load(dataname, fold, aggf, rating, ll, ordering, topk)
        if len(data) == 0 and ll == "no_spn":
            data = load(dataname, fold, "mean", rating, ll, ordering, topk)
            if valnames == ["counterfactual_ll"]:
                ce_vecs = []
                for fid, CEs in data.items():
                    for ce in CEs.values():
                        if "counterfactual" in ce:
                            ce_vecs.append(ce["counterfactual"])
                        else:
                            fails += 1
                vals += eval_sample_lls(ce_vecs, dataname, fold, aggf, rating)
                continue
        for fid, CEs in data.items():
            for ce in CEs.values():
                if all(valname in ce for valname in valnames):
                    if len(valnames) == 1:
                        vals.append(ce[valnames[0]])
                    else:
                        vals.append([ce[valname] for valname in valnames])
                else:
                    fails += 1
    return vals, fails


spns_cache = {}

def eval_sample_lls(samples, data_name, fold, group_f, rating):
    path = f"data_splits/{data_name}/{fold}/models/spn_{group_f}_{rating}.pickle"
    if path in spns_cache:
        spn, groups = spns_cache[path]
    try:
        with open(path, "rb") as f:
            spn, _ = pickle.load(f)
        with open(f"data_splits/{data_name}/{fold}/groups.pickle", "rb") as f:
            (groups, _) = pickle.load(f)
        with open(f"data_splits/{data_name}/{fold}/items_{rating}.pickle", "rb") as f:
            item_map = pickle.load(f)
        groups = [np.array([item_map[item] for item in items]) for items in groups]
        spns_cache[path] = spn, groups
    except FileNotFoundError:
        print("No SPN found")
        return np.nan
    
    lls = []
    for sample in samples:
        cf = np.array(sample)
        if group_f == "sum":
            spn_cf = [cf[g].sum() for g in groups]
        elif group_f == "mean":
            spn_cf = [cf[g].mean() for g in groups]
        elif group_f == "disjunction":
            spn_cf = [(cf[g].sum() > 0).astype(int) for g in groups]
        lls.append(spn.compute_ll(np.array(spn_cf)))
    return lls

vals, fails = get_val_for_confs("time_solving")
print(np.median(vals))

for data in ["amazon", "yelp", "netflix"]:
    print(data)               
    for nth in ["nth", "score"]:
        for ll in ["0.1", "median", "no_spn"]:
            for rat in ["binary", "rating"]:
                for topk in [[5, 10]]:
                    all_fs = ["disjunction", "sum", "mean"]
                    if ll == "no_spn" or rat == "rating":
                        all_fs = ["mean"]
                    for af in all_fs:
                        vals, fails = get_val_for_confs("counterfactual", dataname=[data], aggf=[af], rating=[rat], ll=[ll], topk=topk, ordering=[nth])
                        print(f"{len(vals) / (len(vals) + fails):.2f}", end = " & ")
        print("\\\\")


val = ["timeout","counterfactual"]

vals, fails = get_val_for_confs(val)
if len(vals) != 0:
    touts, ces = [list(v) for v in zip(*vals)]
else:
    touts, ces = [], []

print(f"{np.mean(touts):.3f}", end = " & ")

val = ["timeout","counterfactual"]

for ll in ["0.1", "median", "no_spn"]:
    all_touts, _ = get_val_for_confs("timeout", ll=[ll])
    vals, fails = get_val_for_confs(val, ll=[ll])
    if len(vals) != 0:
        touts, ces = [list(v) for v in zip(*vals)]
    else:
        touts, ces = [], []

    print(f"{(fails - np.sum(all_touts) + np.sum(touts)) / len(all_touts):.2f}", end = " & ")
    print(f"{(fails - np.sum(all_touts) + np.sum(touts)) / fails:.2f}", end = " & ")
    print(" \\\\")

for data in ["amazon", "yelp", "netflix"]:
    print(data)               
    for nth in ["nth", "score"]:
        print(f" & {nth} & ", end="")
        for val in ["counterfactual_ll", "distance"]:
            for rat in ["binary", "rating"]:
                for ll in (["0.1", "median", "no_spn"] if rat == "binary" else ["0.1", "no_spn"]):
                    for topk in [[5, 10]]:
                        af = "disjunction" if rat == "binary" else "mean"
                        vals, fails = get_val_for_confs(val, dataname=[data], aggf=[af], rating=[rat], ll=[ll], topk=topk, ordering=[nth])
                        if np.mean(vals) > 300:
                            print(f"${np.mean(vals):.0f} \\pm {np.std(vals):.0f}$", end = " & ")
                        else:
                            print(f"${np.mean(vals):.2f} \\pm {np.std(vals):.2f}$", end = " & ")
        print(" \\\\")
